adding methods delete cart item, update quantity, view all items, with editing the creation of an order. then handling the return from cart to products table using task scheduling.

+++ routes/api.php
@@ -60,6 +60,7 @@ Route::post('register',[ApiAuthController::class,'register']);
 Route::post('login',[ApiAuthController::class,'login']);
 
 
+
 //get all categories api
 //http://127.0.0.1:8000/api/categories
 Route::get('categories',[ApiCategoryController::class,'all']);
@@ -82,24 +83,24 @@ Route::get('proudcts/show/{id}',[ApiProudctController::class,'show']);
 Route::get('products/search',[ApiProudctController::class,'search'] )->name('apiProductsSearch');
 
 
-Route::controller(CartController::class)->prefix('/carts')->group(function(){
-// Route::middleware('auth:sanctum')->controller(CartController::class)->prefix('/carts')->group(function(){
+// Route::controller(CartController::class)->prefix('/carts')->group(function(){
+Route::middleware('api.auth')->controller(CartController::class)->prefix('carts')->group(function(){
     // APIs for cart
     // 1) add item to cart
     // http://127.0.0.1:8000/api/carts/add
     Route::post('/add', 'addToCart');
 
     // 2) retrive cart items
-    // http://127.0.0.1:8000/api/cart/
+    // http://127.0.0.1:8000/api/carts/{id}
     Route::get('/{id}', 'getCart');
 
     // 3) update quantity of specific item in the cart
-    // http://127.0.0.1:8000/api/cart/id
-    Route::put('/{id}', 'updateCart');
+    // http://127.0.0.1:8000/api/carts/edit/id
+    Route::post('/edit/{id}', 'updateCart');
 
     // 3) delete item in the cart
     // http://127.0.0.1:8000/api/cart/id
-    Route::delete('/{id}', 'deleteCartItem');
+    Route::delete('/{user_id}/{id}', 'deleteCartItem');
 });
